International Debt Statistics
This is a guided project I completed from DataCamp which contains the SQL codes I've written in order to complete the project. The main purpose of the project is for me to be able to demonstrate that I can write SQL queries in order to extract meaningful data for analytic purposes.